chore: context에서 상태 관리 로직 useVendingMachine으로 분리
